Improve structure - Add Path, Hash, Query types - Update dependencies
