+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "svelte-micro",
-  "version": "2.4.0",
+  "version": "2.5.0",
   "description": "Light & reactive router for Svelte",
   "author": "ayndqy",
   "license": "MIT",
@@ -33,8 +33,8 @@
     "svelte": "^3.54.0 || ^4.0.0"
   },
   "devDependencies": {
-    "@sveltejs/kit": "^1.21.0",
-    "@sveltejs/package": "^2.1.0",
+    "@sveltejs/kit": "^1.22.0",
+    "@sveltejs/package": "^2.2.0",
     "svelte-check": "^3.4.0",
     "tslib": "^2.6.0",
     "typescript": "^5.1.0"
