+++ src/index.ts
@@ -1,7 +1,9 @@
-export { type Router, router } from './lib/router'
-export { type Options, type OptionsStore, options } from './lib/options'
-export { type PathStore, type QueryStore, type HashStore, path, query, hash } from './lib/location'
-export { default as Route } from './lib/Route/Route.svelte'
-export { default as Link } from './lib/Link/Link.svelte'
-export { type LinkHandle, linkHandle } from './lib/Link/linkHandle'
-export { type GetPathSegments, getPathSegments } from './lib/getPathSegments'
+export { router, type Router } from './lib/router'
+export { options, type OptionsStore, type Options } from './lib/options'
+export { path, type PathStore, type Path } from './lib/location'
+export { query, type QueryStore, type Query } from './lib/location'
+export { hash, type HashStore, type Hash } from './lib/location'
+export { default as Route } from './lib/components/Route.svelte'
+export { default as Link } from './lib/components/Link.svelte'
+export { linkHandle, type LinkHandle } from './lib/utils/linkHandle'
+export { getPathSegments, type GetPathSegments } from './lib/utils/getPathSegments'
