+++ src/lib/router.ts
@@ -1,7 +1,7 @@
 import { get } from 'svelte/store'
-import { type Options, options } from './options'
+import { options } from './options'
 
-const dispatchLocationChange = (mode: Options['mode']) => {
+const dispatchLocationChange = (mode = get(options).mode) => {
   let type: string = 'popstate'
 
   if (mode === 'window') type = 'popstate'
@@ -16,23 +16,19 @@ export type Router = {
   replace: (url?: string | URL | null, state?: any) => void
 }
 
-export const go: Router['go'] = (delta: number = 0) => {
-  history.go(delta)
-  dispatchLocationChange(get(options).mode)
-}
-
-export const push: Router['push'] = (url, state = null) => {
-  history.pushState(state, '', url)
-  dispatchLocationChange(get(options).mode)
-}
-
-export const replace: Router['replace'] = (url, state = null) => {
-  history.replaceState(state, '', url)
-  dispatchLocationChange(get(options).mode)
-}
-
 export const router: Router = {
-  go: go,
-  push: push,
-  replace: replace,
+  go: (delta: number = 0) => {
+    history.go(delta)
+    dispatchLocationChange()
+  },
+
+  push: (url, state = null) => {
+    history.pushState(state, '', url)
+    dispatchLocationChange()
+  },
+
+  replace: (url, state = null) => {
+    history.replaceState(state, '', url)
+    dispatchLocationChange()
+  },
 }
