+++ src/lib/utils/getPathSegments.ts
