+++ src/lib/utils/getPathWithoutBase.ts
@@ -2,6 +2,5 @@ export type GetPathWithoutBase = (path: string, basePath: string | null) => stri
 
 export const getPathWithoutBase: GetPathWithoutBase = (path, basePath) => {
   if (basePath === null) return path
-
   return path.startsWith(basePath) ? path.slice(basePath.length) : path
 }
